--- packages/utils/src/crypto/header/mac.ts
@@ -1,15 +1,0 @@
-import { createHmac } from 'crypto'
-import { HASH_ALGORITHM, HASH_LENGTH, SECRET_LENGTH } from './constants'
-import { expand } from 'futoin-hkdf'
-
-const HASH_KEY_HMAC = 'HASH_KEY_HMAC'
-
-export function createMAC(secret: Uint8Array, msg: Uint8Array) {
-  if (secret.length != SECRET_LENGTH) {
-    throw Error(`Invalid arguments`)
-  }
-
-  const key = expand(HASH_ALGORITHM, HASH_LENGTH, Buffer.from(secret), SECRET_LENGTH, HASH_KEY_HMAC)
-
-  return createHmac(HASH_ALGORITHM, key).update(msg).digest()
-}
