+++ packages/utils/src/crypto/index.ts
@@ -1,5 +1,5 @@
 export * from './hashIterator'
-export * from './header'
+export * from './packet'
 export * from './por'
 export * from './prg'
 export * from './prp'
